% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_ndvi.R
\encoding{UTF-8}
\name{acceso_ndvi}
\alias{acceso_ndvi}
\title{Calculo del NDVI seccion censal ponderado}
\usage{
acceso_ndvi(
  dir_img,
  sc,
  id_sc,
  catastro,
  radio = 300,
  calculo = mean,
  dev_raster = FALSE,
  landsat8 = FALSE
)
}
\arguments{
\item{dir_img}{Carácter: Directorio donde se encuentra la imagen Landsat, tiene que estar descomprimida.}

\item{sc}{Objeto espacial \code{(SpatialPolygonsDataFrame / sf):} con las geometrías poligonales de
las secciones censales.}

\item{id_sc}{Carácter: nombre de la columna que contiene el código identificador único de cada
sección censal.}

\item{catastro}{Objeto espacial \code{(SpatialPolygonsDataFrame / sf):} con las geometrías poligonales
de la capa 'building' de la base de datos espacial del catastro.}

\item{radio}{\code{Numérico:} Valor del radio del buffer para la intersección entre los centroides
de los polígonos catastrales y capa urban atlas, expresado en metros. Por defecto toma valor de 300m.}

\item{calculo}{Función para resumir los valores NDVI en la capa de polígonos, por defecto \code{mean}.
Se puede modificar (ej. median, min o max).}

\item{dev_raster}{Lógico: Si es TRUE en el resultado se incluye la capa ráster con los valores del NDVI.}

\item{landsat8}{Lógico: Por defecto se calculan de manera automática las imagenes de landsat-5
y landsat-7. En caso de querer calcular una imagen landsat-8, cambiar el parámetro a \code{TRUE}.}
}
\value{
El resultado que devuelve es el mismo objeto espacial \code{(SpatialPolygonsDataFrame / sf)}
introducido en el parámetro \code{sc} de la función, con una columna nueva llamada 'ndvi_pond' que tiene el
resultado para cada polígono.

Si también se quiere obtener el mapa ráster con el resultado del ndvi, se tiene que cambiar el
parámetro \code{dev_raster = TRUE}. Entonces el resultado devuelto es una lista con dos elementos:
El primero es el objeto espacial \code{(SpatialPolygonsDataFrame / sf)} comentado anteriormente y
el segundo una \code{RasterLayer} con el resultado del ndvi en forma de mapa de píxeles.
}
\description{
Extrae el valor del índice de vegetación normalizado (NDVI) de cada sección censal
ponderado. Para ello utiliza los datos de una imagen del satelite Landsat, a la que se le realiza
el cálculo del NDVI, una capa de polígonos correspondiente a las secciones censales y la base de
datos del catastro. Por defecto el método de extracción es la media de píxeles de cada polígono,
este método se puede modificar con el parámetro \code{calculo}.
}
\details{
Esta función es una variante más compleja de \link[greennessr]{ndvi}, teniendo
en cuenta la accesibilidad a espacios verdes urbanos de los habitantes de una sección censal en
función de la ubicación de su edificio. En primer lugar la se extrae el centroide de cada
edificio del mapa catastral,seguidamente se genera un buffer de 300 metros (modificable con
el parámetro \code{radio}) que intersecta con el mapa ráster resultante del cálculo del NDVI
y se calcula el valor medio de NDVI comprendidos dentro del buffer.Finalmente se calcula la
media de los niveles de NDVI de los edificio de cada sección censal ponderando por el número
de viviendas que se albergan en cada edificio.

Para el uso de la función se necesita descargar los edificios de la base de datos catastral (capa
BUILDING). Estos están disponible a través del servicio de
\href{http://www.catastro.minhap.es/INSPIRE/buildings/ES.SDGC.BU.atom.xml}{descargas ATOM} del
servicio INSPIRE de la página del catastro. Además también existe un plugin para QGIS, llamado
'Spanish Insipire Catastral Downloader', que proporciona una interfaz más amable para la descarga
de los datos. Es recomendable cargar la capa con extensión '.gml' porque su tiempo de lectura de datos
es menor.

El índice de vegetación normalizado (NDVI) se utiliza para estimar la cantidad, calidad y el
desarrollo de la vegetación. Este estimador se calcula mediante los valores de intensidad de la radiación
que las plantas emiten o reflejan en ciertos rangos del espectro electromagnético. Concretamente en las
longitudes de onda del rojo y del infrarrojo cercano (nir). El cálculo es el siguiente:

NDVI = (nir - rojo) / (nir + rojo)

Para que la función pueda calcularlo es necesario una imagen satélite multiespectral, en este caso se
utilizan las de los satélites Landsat-5, Landsat-7 y Landsat-8 (en caso de utilizar Landsat-8 cambiar
el parámetro \code{landsat8 = TRUE}). Para descargar las imágenes es recomendable usar la página web
\href{https://earthexplorer.usgs.gov/}{earth explorer de la USGS.} Es muy importante que al realizar
la descarga se tenga en cuenta lo siguiente:

\itemize{
\item Que las imágenes seleccionadas sean de la colección 'Landsat Collection 1 Level-1 (L1TP)'
ya que están radiometricamente calibradas y ortorectificadas.

\item Que la zona de estudio no este cubierta por nubes, en la página web de
\href{https://earthexplorer.usgs.gov/}{earth explorer} podrá visualizar la imagen antes de descargarla.

\item Que la imagen cubra toda la superficie de la zona de estudio. En el caso de tener que descargar
más de una imagen para cubrir su zona de estudio tendrá que juntar las imágenes previamente antes
de introducirlas en la función. Esto lo puede realizar usando la función \code{\link[raster]{merge}}.
}
En este \href{https://earthexplorer.usgs.gov/documents/helptutorial.pdf}{link} encontrará más detalles a cerca
de la descarga y los distintos niveles de procesamiento de las imágenes.
No hace falta tener en cuenta los sistemas de referencia de los datos geográficos, la propia función
los cambiará si fuese necesario.
}
\examples{
\dontrun{
library(greennessr)
sc_cordoba <- acceso_ndvi(
  dir_img = '~01_data/landsat/Cordoba',
  sc = cordoba_sc,
  id_sc = "seccion",
  catastro = cordoba_catastro,
  radio = 300,
  calculo = mean,
  dev_raster = TRUE,
  landsat8 = FALSE
)
sc_cordoba}

}
