% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_urban_green.R
\encoding{UTF-8}
\name{acceso_verde}
\alias{acceso_verde}
\title{Accesibilidad a areas verdes}
\usage{
acceso_verde(
  sc,
  id_sc,
  urban,
  catastro,
  radio = 300,
  categorias = c(14100, 14200)
)
}
\arguments{
\item{sc}{Objeto espacial \code{(SpatialPolygonsDataFrame / sf):} con las geometrías poligonales
de las secciones censales.}

\item{id_sc}{Carácter: nombre de la columna que contiene el código identificador único de cada
sección censal.}

\item{urban}{Objeto espacial \code{(SpatialPolygonsDataFrame / sf):} con las geometrías poligonales
del mapa urban atlas.}

\item{catastro}{Objeto espacial \code{(SpatialPolygonsDataFrame / sf):} con las geometrías poligonales
de la capa 'building' de la base de datos espacial del catastro.}

\item{radio}{\code{Numérico:} Valor del radio del buffer para la intersección entre los centroides
de los polígonos catastrales y capa urban atlas, expresado en metros. Por defecto toma valor de 300m.}

\item{categorias}{\code{Vector:} Correspondiente a las categorías de espacios verdes de la base de datos
del urban atlas.}
}
\value{
El resultado que devuelve es el mismo objeto espacial \code{(SpatialPolygonsDataFrame / sf)}
introducido en el parámetro \code{sc} de la función, con una columna nueva llamada accs_verde que
tiene el resultado para cada polígono. Este valor está comprendido entre 0 y 1. Un valor 1 quiere decir
que la media de los buffers alrededor de los centroides de los edificios de cada sección censal están
integramente cubiertos por espacios verdes y un valor 0 que no lo están. Se tiene que tener en cuenta
que en los edificios que comprenden un número alto de viviendas obtienen un mayor peso en la media.
}
\description{
Extrae el índice de accesibilidad ponderado de áreas verdes urbanas correspondiente a cada
sección censal. Para ello utiliza los datos del Urban Atlas, el mapa de secciones censales y la base de
datos del catastro. Por defecto la función tiene definidas las categorías de espacios verdes
correspondientes a 'Green urban areas' y Sports and leisure facilities', pero estas se pueden modificar
con el parámetro \code{categorias}.
}
\details{
Esta función es una variante más compleja de \link[greennessr]{area_verde}, que tiene en cuenta
la accesibilidad a espacios verdes urbanos de los habitantes de una sección censal en función de la
ubicación de su edificio. En primer lugar la se extrae el centroide de cada edificio del mapa catastral,
seguidamente se genera un buffer de 300 metros (modificable con el parámetro \code{radio}) que
intersecta con las áreas verdes urbanas y se calcula la superficie que ocupa está área respecto el
buffer. Finalmente se calcula la media de las áreas de los edificio de cada sección censal ponderando
por el número de viviendas que se albergan en cada edificio.

Para el uso de la función se necesita descargar los edificios de la base de datos catastral (capa
BUILDING). Estos están disponible a través del servicio de
\href{http://www.catastro.minhap.es/INSPIRE/buildings/ES.SDGC.BU.atom.xml}{descargas ATOM} del
servicio INSPIRE de la página del catastro. Además también existe un plugin para QGIS, llamado
'Spanish Insipire Catastral Downloader', que proporciona una interfaz más amable para la descarga
de los datos. Es recomendable cargar la capa con extensión '.gml' porque su tiempo de lectura de datos
es menor.

También es imprescindible descargar los datos del Urban atlas. Este es un mapa
de cubiertas y usos del suelo de las principales ciudades europeas creado por el proyecto Copernicus,
consta de dos ediciones 2006 y 2012. Para el primer año solo están disponibles las principales
ciudades españolas, en la versión de 2012 aumenta considerablemente el número de ciudades.
\href{https://land.copernicus.eu/local/urban-atlas}{Clic para más información y descargas}.

No hace falta tener en cuenta los sistemas de referencia de los datos geográficos, la propia función
los cambiará si fuese necesario.
}
\examples{
library(greennessr)

# carga de datos de ejemplo
data("catastro_castellon")
data("secciones_castellon")
data("urban_atlas_castellon")

# comprobar que son obejetos espaciales
class(catastro_castellon)
class(secciones_castellon)
class(urban_atlas_castellon)

head(secciones_castellon) # 'secccion'

accs_verde <- acceso_verde(
  sc         = secciones_castellon,
  id_sc      = 'seccion',
  urban      = urban_atlas_castellon,
  catastro   = catastro_castellon,
  radio      = 300,
  categorias = c(14100, 14200)
)

summary(accs_verde$accs_verde)

}
