% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{catastro_castellon}
\alias{catastro_castellon}
\title{Edificios de Castello de la Plana}
\format{
Objeto data.frame con extensión espacial sf.
}
\usage{
catastro_castellon
}
\description{
Cartografía de edificios de la base de datos del catastro en el
municipio de Castelló de la Plana para ser usada como ejemplo en las
funciones de \code{acceso_verde y area_verde}.
}
\details{
Los datos están disponible a través del servicio de
\href{http://www.catastro.minhap.es/INSPIRE/buildings/ES.SDGC.BU.atom.xml}{descargas ATOM} del
servicio INSPIRE de la página del catastro. Además también existe un plugin para QGIS, llamado
'Spanish Insipire Catastral Downloader', que proporciona una interfaz más amable para la descarga
de los datos. Es recomendable cargar la capa con extensión '.gml' porque su tiempo de lectura de datos
es menor.
}
\examples{
library(greennessr)
data(catastro_castellon)

}
\keyword{datasets}
