% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{secciones_castellon}
\alias{secciones_castellon}
\title{Secciones censales Castello de la Plana}
\format{
Objeto data.frame con extensión espacial sf.
}
\usage{
secciones_castellon
}
\description{
Cartografía de las secciones censales del municipio de Castelló de la Plana
para ser usada como ejemplo en las funciones de \code{acceso_verde y area_verde}
}
\details{
Datos extraidos de la función \code{carto_medea3} de la librería \code{medear},
está cartografía tiene la particularidad que esta unificada acorde para el desarrollo del
proyecto MEDEA3, para la concordancia de las variaciones del seccionado en el período de
estudio (1996-2015).
}
\examples{
library(greennessr)
data(secciones_castellon)

}
\keyword{datasets}
