% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{urban_atlas_castellon}
\alias{urban_atlas_castellon}
\title{Urban Atlas de Castello de la Plana}
\format{Objeto data.frame con extensión espacial sf.}
\usage{
urban_atlas_castellon
}
\description{
Cartografía de la base de datos Urban Atlas en el municipio de Castelló
de la Plana para ser usada como ejemplo en las funciones de
\code{acceso_verde y area_verde}.
}
\details{
El urban atlas es un mapa de cubiertas y usos del suelo de las principales
ciudades europeas creado por el proyecto Copernicus, consta de dos ediciones 2006
y 2012. Para el primer año solo están disponibles las principales ciudades españolas,
en la versión de 2012 aumenta considerablemente el número de ciudades.
\href{https://land.copernicus.eu/local/urban-atlas}{Clic para más información y descargas}
}
\examples{
library(greennessr)
data(urban_atlas_castellon)

}
\keyword{datasets}
